transforms: (preprocess-mlperf-tiny) remove pass
factor out the transposeconstants to a separate file, this is the only
pattern we still need

formatting
